requirements.txt
aiohttp .
aiosignal .
annotated-types .
anyio .
asgiref .
attrs .
backoff .
bcrypt .
beautifulsoup4 .
bs4 .
build .
cachetools .
certifi .
charset-normalizer .
chroma-hnswlib .
chromadb .
click .
colorama .
coloredlogs .
dataclasses-json .
Deprecated .
distro .
faiss-cpu .
fastapi .
filelock .
flatbuffers .
frozenlist .
fsspec .
google-auth .
googleapis-common-protos .
greenlet .
grpcio .
h11 .
httpcore .
httptools .
httpx .
huggingface-hub .
humanfriendly .
idna .
importlib-metadata .
importlib_resources .
jsonpatch .
jsonpointer .
langchain .
langchain-community .
langchain-core .
langchain-openai .
langchain-text-splitters .
langsmith .
markdown-it-py .
marshmallow .
mdurl .
mmh3 .
monotonic .
mpmath .
multidict .
mypy-extensions .
numpy .
oauthlib .
onnxruntime .
openai .
opentelemetry-api .
opentelemetry-exporter-otlp-proto-common .
opentelemetry-exporter-otlp-proto-grpc .
opentelemetry-instrumentation .
opentelemetry-instrumentation-asgi .
opentelemetry-instrumentation-fastapi .
opentelemetry-proto .
opentelemetry-sdk .
opentelemetry-semantic-conventions .
opentelemetry-util-http .
orjson .
overrides .
packaging .
pip .
posthog .
protobuf .
pulsar-client .
pyasn1 .
pyasn1_modules .
pydantic .
pydantic_core .
Pygments .
PyPika .
pyproject_hooks .
pyreadline3 .
python-dateutil .
python-dotenv .
PyYAM .
regex .
requests .
requests-oauthlib .
rich .
rsa .
setuptools .
shellingham .
six .
sniffio .
soupsieve .
SQLAlchemy .
starlette .
sympy .
tenacity .
tiktoken .
tokenizers .
tqdm .
typer .
typing_extensions .
typing-inspect .
urllib3 .
uvicorn .
watchfiles .
websocket-client .
websockets .
wrapt .
yarl .
zipp .